Separate string solving code
Move string solving code into a separate namespace
Mata::Nfa::StringSolving and hide its implementation in a its own
subfolder src/nfa/string-solving.
