+++ src/nfa/string-solving/tests-nfa-noodlification.cc
@@ -2,12 +2,13 @@
 
 #include <unordered_set>
 
-#include "../3rdparty/catch.hpp"
+#include "catch.hpp"
 
-#include <mata/nfa.hh>
-#include <mata/noodlify.hh>
+#include "mata/nfa.hh"
+#include "mata/nfa-string-solving.hh"
 
 using namespace Mata::Nfa;
+using namespace Mata::Nfa::StringSolving;
 using namespace Mata::util;
 using namespace Mata::Parser;
